Treat CNAME records as 'non strict' for validation purposes
Strict validation requires labels to match the following regex:

[^a-zA-Z0-9-]

However, this does not allow CNAME records to be created which
contain an underscore in the label, even though this is valid and
required by some providers, e.g. validating AWS:

_randomstring 300 CNAME randomstring.randomstring.acm-validations.aws.

The regex itself cannot be changed to allow underscores, as this
would permit entries such as:

_host 10 A 127.0.0.1

Such entries appear to be invalid, and are definitely rejected by
the Mythic Beasts API. They would also fail the dns-sync-zone
test suite.

Since CNAME labels seem to be subject to much laxer rules than
other types, the simplest solution is to make CNAME records 'not
strict', in the same way as TXT and SRV records. This commit
includes that change, as well as an additional test to ensure that
CNAME labels containing underscores pass validation.

Fixes #47