Remove support for old style zone files
As far as I'm aware, no one was using the old style zone files, so
after several months notice (starting on 12 May 2020), I'm removing
support for this format. This simplifies both the zone file parsing
and the argument parsing parts of the code and reduces the amount
of code that needs to be maintained.

A new release will be generated with a major version increment, as
this is a breaking change.

Closes #43