Allow underscores in labels, including at the start
Underscores are permitted at the start of labels, but the validation
was preventing them. The following changes have been made:

Remove 2 tests which considered underscores to be an invalid label.
Add 1 test to ensure labels with underscores are valid.
Change regex to allow underscores in labels.

This does allow records like:

_host 10 A 12.13.14.15

which will be rejected by API. However in this case technically
the API can be seen as incorrect from a DNS perspective, even if
some resolvers will have problems with records starting with an
underscore.